+++ src/emulate.js
@@ -22,78 +22,27 @@
 
 var cordova_util      = require('./util'),
     path              = require('path'),
-    child_process     = require('child_process'),
-    events            = require('./events'),
     hooker            = require('./hooker'),
-    Q                 = require('q'),
-    DEFAULT_OPTIONS   = ["--emulator"],
-    os                = require('os');
-
-// Returns a promise.
-function shell_out_to_emulate(projectRoot, platform, options) {
-    var cmd = path.join(projectRoot, 'platforms', platform, 'cordova', 'run'),
-        args = options.length ? DEFAULT_OPTIONS.concat(options) : DEFAULT_OPTIONS,
-        d = Q.defer(),
-        errors = "",
-        child;
-
-    if (os.platform() === 'win32') {
-        args = ['/c',cmd].concat(args);
-        cmd = 'cmd';
-    }
-
-    events.emit('log', 'Running on emulator for platform "' + platform + '" via command "' + cmd + '" ' + args.join(" "));
-
-    //using spawn instead of exec to avoid errors with stdout on maxBuffer
-    child = child_process.spawn(cmd, args);
-
-    child.stdout.setEncoding('utf8');
-    child.stdout.on('data', function (data) {
-        events.emit('verbose', data);
-    });
-
-    child.stderr.setEncoding('utf8');
-    child.stderr.on('data', function (data) {
-        events.emit('verbose', data);
-        errors = errors + data;
-    });
-
-    child.on('close', function (code) {
-        events.emit('verbose', "child_process.spawn(" + cmd + "," + "[" + args.join(", ") + "]) = " + code);
-        if (code === 0) {
-            events.emit('log', 'Platform "' + platform + '" deployed to emulator.');
-            d.resolve();
-        } else {
-            d.reject(new Error('An error occurred while emulating/deploying the ' + platform + ' project. ' + errors));
-        }
-    });
-
-    return d.promise;
-}
+    superspawn        = require('./superspawn'),
+    Q                 = require('q');
 
 // Returns a promise.
 module.exports = function emulate(options) {
-    var projectRoot = cordova_util.cdProjectRoot(),
-        hooks;
-
-    if (!options) {
-        options = {
-            verbose: false,
-            platforms: [],
-            options: []
-        };
-    }
-
+    var projectRoot = cordova_util.cdProjectRoot();
     options = cordova_util.preProcessOptions(options);
 
-    hooks = new hooker(projectRoot);
+    var hooks = new hooker(projectRoot);
     return hooks.fire('before_emulate', options)
     .then(function() {
         // Run a prepare first!
         return require('../cordova').raw.prepare(options.platforms);
     }).then(function() {
+        // Deploy in parallel (output gets intermixed though...)
         return Q.all(options.platforms.map(function(platform) {
-            return shell_out_to_emulate(projectRoot, platform, options.options);
+            var cmd = path.join(projectRoot, 'platforms', platform, 'cordova', 'run');
+            var args = ['--emulator'].concat(options.options);
+
+            return superspawn.spawn(cmd, args, {stdio: 'inherit', printCommand: true});
         }));
     }).then(function() {
         return hooks.fire('after_emulate', options);